tp-datos-personales/requirements.txt
lxml==4.8.0
numpy==1.22.3
pandas==1.4.2
pykml==0.2.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
